Update drag-to-select behaviour.
This commit changed how drag-to-select behaves. Now if a user simply clicks on an empty area, it doesn't immediately start a new crop region at the user's point of click. But rather, it only takes effect after the user starts dragging. This means that a single click on any empty area now does nothing.