Modify redraw method to improve performance and fix edge cases.
This commit improves the redraw method by leveraging the performance improvements of using requestAnimationFrame. Croppr now only redraws when the browser is able to.

Other changes introduced is dynamically bringing certain handles to the foreground. If the region is in the lower right quadrant, the top left handle is brought forward. If the region is in the upper left region quadrant, the bottom right handle is brought forward. And so on and forth. This prevents edge cases (when aspect ratio mode is on) where the crop region becomes unresizable because all the handles overlap one another, and the uppermost handle resizes towards a direction that has already hit the region boundary.