Fix a moving behaviour bug when user clicks on an empty area beforehand.
When a user clicks on an empty area, Croppr thinks you're trying to drag a new crop region from the click point. For a better user experience, Croppr was made in a way that detects if it's merely an accidental click on an empty area and retains the existing crop region (meaning the drag a new crop region process only begins when the user actually drags). Unfortunately, releases prior to this commit failed detecting this. This commit fixes that and the detection works perfectly now.